chore : resp 타입 수정

+++ src/main/java/org/knulikelion/challengers_backend/service/Impl/ClubJoinServiceImpl.java
@@ -36,7 +36,7 @@ public class ClubJoinServiceImpl implements ClubJoinService {
     private final ClubService clubService;
 
     @Override
-    public BaseResponseDto createJoinRequest(String token, Long clubId, String comment) {
+    public ResponseEntity<BaseResponseDto> createJoinRequest(String token, Long clubId, String comment) {
 
         Club club = clubRepository.findById(clubId).orElseThrow(ClubNotFoundException::new);
         User user = userRepository.findByEmail(jwtTokenProvider.getUserEmail(token));
@@ -44,10 +44,10 @@ public class ClubJoinServiceImpl implements ClubJoinService {
         List<Club> foundClub = clubRepository.findAllByClubManager(user);
         for(Club temp : foundClub) {
             if(temp.getId().equals(clubId)) {
-                return BaseResponseDto.builder()
+                return new ResponseEntity<>(BaseResponseDto.builder()
                         .success(false)
                         .msg("클럽 생성자는 가입을 요청할 수 없습니다.")
-                        .build();
+                        .build(),HttpStatus.BAD_REQUEST);
             }
         }
 
@@ -55,11 +55,18 @@ public class ClubJoinServiceImpl implements ClubJoinService {
 
         for(UserClub userClub : userClubList) {
             if(userClub.getClub().getId().equals(clubId)) {
-                return BaseResponseDto.builder()
+                return new ResponseEntity<>(BaseResponseDto.builder()
                         .success(false)
                         .msg("이미 가입된 클럽입니다.")
-                        .build();
+                        .build(),HttpStatus.BAD_REQUEST);
             }
+        }
+
+        if(!club.isClubApproved()){
+            return new ResponseEntity<>(BaseResponseDto.builder()
+                    .success(false)
+                    .msg("승인된 클럽이 아닙니다. 관리자에게 문의해 주세요.")
+                    .build(),HttpStatus.UNAUTHORIZED);
         }
 
         if (user == null) {
@@ -73,15 +80,15 @@ public class ClubJoinServiceImpl implements ClubJoinService {
 
             clubJoinRepository.save(clubJoin);
 
-            return BaseResponseDto.builder()
+            return ResponseEntity.ok(BaseResponseDto.builder()
                     .success(true)
                     .msg("클럽 가입 요청 완료.")
-                    .build();
+                    .build());
         } else {
-            return BaseResponseDto.builder()
+            return ResponseEntity.ok(BaseResponseDto.builder()
                     .success(false)
                     .msg("중복된 요청입니다.")
-                    .build();
+                    .build());
         }
     }
 
@@ -91,7 +98,7 @@ public class ClubJoinServiceImpl implements ClubJoinService {
         Club club = clubRepository.getById(clubId);
 
 
-        if (!club.getClubManager().equals(user) || club.getClubApproved()!=1) {
+        if (!club.getClubManager().equals(user) || !club.isClubApproved()) {
             return new ResponseEntity<>(BaseResponseDto.builder()
                     .success(false)
                     .msg("클럽 멤버 수락 권한이 없습니다.")
@@ -111,25 +118,35 @@ public class ClubJoinServiceImpl implements ClubJoinService {
     }
 
     @Override
-    public BaseResponseDto rejectJoinRequest(Long clubId, String userEmail, String rejectUserEmail) {
+    public ResponseEntity<BaseResponseDto> rejectJoinRequest(Long clubId, String userEmail, String rejectUserEmail) {
         User user = userRepository.findByEmail(userEmail);
         User findUser = userRepository.findByEmail(rejectUserEmail);
         Club club = clubRepository.findById(clubId).orElseThrow(ClubNotFoundException::new);
 
-        BaseResponseDto baseResponseDto = new BaseResponseDto();
+        if(!club.isClubApproved()){
+            return new ResponseEntity<>(BaseResponseDto.builder()
+                    .success(false)
+                    .msg("승인되지 않은 클럽입니다. 관리자에게 문의해주세요.")
+                    .build(),HttpStatus.UNAUTHORIZED);
+        }
 
         if (!club.getClubManager().equals(user)) {
-            baseResponseDto.setSuccess(false);
-            baseResponseDto.setMsg("클럽 멤버 거절 권한이 없습니다.");
+            return new ResponseEntity<>(BaseResponseDto.builder()
+                    .success(false)
+                    .msg("클럽 멤버 거절 권한이 없습니다. 관리자에게 문의해주세요.")
+                    .build(),HttpStatus.UNAUTHORIZED);
         } else {
             ClubJoin clubJoin = clubJoinRepository.findByClubIdAndUserId(clubId, findUser.getId());
             clubJoin.setClub(null);
             clubJoin.setUser(null);
             clubJoinRepository.delete(clubJoin);
-            baseResponseDto.setSuccess(true);
-            baseResponseDto.setMsg("클럽 멤버 합류를 거절하였습니다.");
+
+            return ResponseEntity.ok(BaseResponseDto.builder()
+                    .success(false)
+                    .msg("클럽 멤버 함류를 거절하였습니다.")
+                    .build());
         }
-        return baseResponseDto;
+
     }
 
     @Override
